+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `blake3` feature, implementing `Aggregate` for `blake3::Hash` [#11]
+
 <!-- ISSUES -->
+[#11]: https://github.com/dusk-network/merkle/issues/11
 
 <!-- VERSIONS -->
 [Unreleased]: https://github.com/dusk-network/merkle/compare/v0.1.0...HEAD